+++ tests-interactive/opengl-shaderpipeline-with-api/main.cpp
@@ -46,8 +46,8 @@ public:
         if (state.getName() == "FrameStart") {
             float time = state.getValue();
             // Calculate model matrix based on time
-            VRMatrix4 modelMatrix = VRMatrix4::rotationX(0.5*time);
-            modelMatrix = modelMatrix * VRMatrix4::rotationY(0.5*time);
+            VRMatrix4 modelMatrix = VRMatrix4::rotationX(0.5f*time);
+            modelMatrix = modelMatrix * VRMatrix4::rotationY(0.5f*time);
             for (int f = 0; f < 16; f++) {
                 model[f] = modelMatrix.getArray()[f];
             }
@@ -253,7 +253,7 @@ public:
 	/// Compiles shader
 	GLuint compileShader(const std::string& shaderText, GLuint shaderType) {
 		const char* source = shaderText.c_str();
-		int length = shaderText.size();
+		int length = (int)shaderText.size();
 		GLuint shader = glCreateShader(shaderType);
 		glShaderSource(shader, 1, &source, &length);
 		glCompileShader(shader);
